+++ lib/router_management.dart
@@ -1,12 +1,11 @@
 library router_management;
 
-export 'src/data/mixins/navigation.dart';
-export 'src/data/models/navigation_page.dart';
-export 'src/data/models/page_arguments.dart';
-export 'src/data/models/page_settings.dart';
-export 'src/data/validators/page_validator.dart';
-export 'src/presentation/extensions/navigation_extension.dart';
-export 'src/presentation/extensions/page_arguments_extension.dart';
-export 'src/presentation/services/navigation_service.dart';
-export 'src/presentation/ui/navigation_router.dart';
-export 'src/presentation/widgets/page_widget.dart';
+export 'src/extensions/navigation_extension.dart';
+export 'src/extensions/page_arguments_extension.dart';
+export 'src/mixins/navigation.dart';
+export 'src/models/navigation_page.dart';
+export 'src/models/page_arguments.dart';
+export 'src/services/navigation_service.dart';
+export 'src/ui/navigation_router.dart';
+export 'src/ui/page_settings.dart';
+export 'src/validators/page_validator.dart';