+++ lib/src/extensions/navigation_extension.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/widgets.dart';
-
-import '../../data/mixins/navigation.dart';
-import '../services/navigation_service.dart';
+import 'package:router_management/src/mixins/navigation.dart';
+import 'package:router_management/src/services/navigation_service.dart';
 
 /// [NavigationExtension] is used to get the current [Navigation] instance on
 /// the [BuildContext]