+++ lib/src/extensions/page_arguments_extension.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/widgets.dart';
-
-import '../../data/models/page_arguments.dart';
+import 'package:router_management/src/models/page_arguments.dart';
 
 /// [PageArgumentsExtension] is used to get the current [PageArguments] on the
 /// [BuildContext]