+++ lib/src/mixins/navigation.dart
@@ -1,4 +1,4 @@
-import '../models/page_settings.dart';
+import 'package:router_management/src/ui/page_settings.dart';
 
 /// [Navigation] is the core navigation interface and is used to controls the
 /// navigator 2.0