+++ lib/src/models/navigation_page.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/widgets.dart';
-
-import '../../core/exceptions/navigation_exception.dart';
-import '../mixins/navigation.dart';
-import '../validators/page_validator.dart';
-import 'page_arguments.dart';
+import 'package:router_management/src/exceptions/navigation_exception.dart';
+import 'package:router_management/src/mixins/navigation.dart';
+import 'package:router_management/src/models/page_arguments.dart';
+import 'package:router_management/src/validators/page_validator.dart';
 
 /// [NavigationPage] is used to create a page into the [NavigationRouter]
 class NavigationPage {