+++ lib/src/validators/page_validator.dart
@@ -1,6 +1,6 @@
-import '../mixins/navigation.dart';
-import '../models/navigation_page.dart';
-import '../models/page_arguments.dart';
+import 'package:router_management/src/mixins/navigation.dart';
+import 'package:router_management/src/models/navigation_page.dart';
+import 'package:router_management/src/models/page_arguments.dart';
 
 /// [PageValidator] can be used to validate any pages with
 /// [NavigationPage.validators]